refactor: cleanunp no more needed code
